system service: split out cgroups call into linux specific file
So that we do not cause compile errors on freebsd.
