+++ cmd/podman/system/service_abi_common.go
@@ -5,3 +5,6 @@ package system
 // Currently, we only need servicereaper on Linux to support slirp4netns.
 func maybeStartServiceReaper() {
 }
+
+// There is no cgroup on non linux.
+func maybeMoveToSubCgroup() {}
