+++ pkg/domain/infra/abi/system_freebsd.go
@@ -1,4 +1,13 @@
 package abi
 
+import (
+	"context"
+)
+
 // Default path for system runtime state
 const defaultRunPath = "/var/run"
+
+// SetupRootless in a NOP for freebsd as it only configures the rootless userns on linux.
+func (ic *ContainerEngine) SetupRootless(_ context.Context, noMoveProcess bool) error {
+	return nil
+}
